@@ -1,17 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="AngleSharp" Version="0.17.1" />
-		<PackageReference Include="AngleSharp.Css" Version="0.16.4" />
+		<PackageReference Include="AngleSharp" Version="1.1.2" />
+		<PackageReference Include="AngleSharp.Css" Version="0.17.0" />
 		<!--PackageReference Include="Selenium.RC" Version="3.1.0" /-->
-		<PackageReference Include="Selenium.Support" Version="4.6.0" />
-		<PackageReference Include="Selenium.WebDriver" Version="4.6.0" />
-		<PackageReference Include="Selenium.WebDriver.ChromeDriver" Version="107.0.5304.6200" />
+		<PackageReference Include="Selenium.Support" Version="4.25.0" />
+		<PackageReference Include="Selenium.WebDriver" Version="4.25.0" />
+		<PackageReference Include="Selenium.WebDriver.ChromeDriver" Version="129.0.6668.5800" />
 		<PackageReference Include="Yaap.zhipwang" Version="0.6.6" />
 	</ItemGroup>
 
