@@ -9,15 +9,11 @@ using Microsoft.AspNetCore.Mvc;
 using Oocw.Database.Models;
 using Oocw.Backend.Controllers;
 using Oocw.Backend.Models;
+using Oocw.Backend.Services;
+using Microsoft.AspNetCore.Http;
 
-namespace Oocw.Backend.Services;
+namespace Oocw.Backend.Auth;
 
-public class JwtConfig
-{
-    public string Secret { get; set; } = "RCDWRCDWRCDWRCDW";
-    public int AccessExpiration { get; set; } = 1;
-    public int RefreshExpiration { get; set; } = 60;
-}
 
 public static class AuthUtils
 {
@@ -25,6 +21,7 @@ public static class AuthUtils
 
     public const string KEY_UPDATED_AT = "updated_at";
     public const string KEY_USAGE = "usage";
+    public const string KEY_ITEM_USER = "user";
 
     public static string GenerateRefreshToken(this User user, JwtConfig config)
     {
@@ -32,12 +29,12 @@ public static class AuthUtils
 
         var tokenDescriptor = new SecurityTokenDescriptor
         {
-            Subject = new ClaimsIdentity(new[]
-            {
+            Subject = new ClaimsIdentity(
+            [
                 new Claim(JwtRegisteredClaimNames.Aud, user.Id.ToString()),
-                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()), 
+                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
                 new Claim(KEY_UPDATED_AT, user.RefreshTime.ToBinary().ToString())
-            }),
+            ]),
 
             Expires = DateTime.UtcNow.AddDays(config.RefreshExpiration),
             SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
@@ -56,12 +53,12 @@ public static class AuthUtils
 
         var tokenDescriptor = new SecurityTokenDescriptor
         {
-            Subject = new ClaimsIdentity(new[]
-            {
+            Subject = new ClaimsIdentity(
+            [
                 new Claim(JwtRegisteredClaimNames.Aud, user.Id.ToString()),
-                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()), 
+                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
                 new Claim(KEY_USAGE, accessType)
-            }),
+            ]),
 
             Expires = DateTime.UtcNow.AddMinutes(config.AccessExpiration),
             SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
@@ -81,7 +78,7 @@ public static class AuthUtils
             ValidateIssuerSigningKey = true,
             IssuerSigningKey = new SymmetricSecurityKey(key),
             ValidateIssuer = false,
-            ValidateAudience = false, 
+            ValidateAudience = false,
         };
 
         try
@@ -126,32 +123,7 @@ public static class AuthUtils
         }
     }
 
-    public static ActionResult<T> RequireAuth<T>(
-        this ServedController ctrl, 
-        Func<User, ActionResult<T>> callback, 
-        string? accessType = "")
-    {
-        // verify access token
-        var verified = false;
-        var fullAccessType = string.IsNullOrEmpty(accessType) ? Definitions.KEY_ACCESS_TOKEN : string.Format(Definitions.KEY_ACCESS_TOKEN_FORMAT, accessType);
-        User? user = null;
-        var hasAccess = ctrl.Request.Cookies.TryGetValue(fullAccessType, out var tokenRaw);
-        if (hasAccess && VerifyAccessToken(tokenRaw ?? "", ctrl.JwtConfig, accessType, ctrl.DBService, out user))
-            verified = true;
-
-        // if failed, varify refresh token and grant access token if possible
-        if (!verified)
-        {
-            var hasRefresh = ctrl.Request.Cookies.TryGetValue(Definitions.KEY_REFRESH_TOKEN, out var refreshTokenRaw);
-            if (hasRefresh && VerifyRefreshToken(refreshTokenRaw ?? "", ctrl.JwtConfig, ctrl.DBService, out user)) 
-            {
-                verified = true;
-                // TODO check if possible to grant access token
-                // it is always possible at this time
-                ctrl.Response.Cookies.Append(fullAccessType, GenerateAccessToken(user!, accessType, ctrl.JwtConfig));
-            }
-        }
-
-        return verified ? callback(user!) : ctrl.Unauthorized(new StandardResult(Definitions.CODE_ERR_AUTH_FAILED));
+    public static User? GetUser(this HttpContext context) {
+        return context.Items[KEY_ITEM_USER] as User;
     }
 }
